refactor(table): move hooks to table/hooks.rs; fix clippy (collapsible_if, ptr_arg)
