+++ vantage-table/src/table/mod.rs
@@ -18,6 +18,9 @@
 pub mod base;
 pub use base::*;
 
+pub mod hooks;
+pub use hooks::*;
+
 pub mod impls;
 pub use impls::*;
